Commit some work-in-progress on typed struct gen.
This has been on the clipboard for a while now.  When I first wrote it,
the idea that we'd need multiple NodeBuilders per typed Node seemed a
little staggering; now that the idea has had a while to cook, it seems
clear and uncontrovertial, so this diff now looks like far less of a
noodle-baker than it originally was.
(I had been planning to switch to working on runtime (that is,
wrapper-based, non-codegen) typed nodes for a while to validate this
idea, and now I have some partial diffs from that to land too, but
the validation returned "true", so now I regret having multiple irons
in the fire... sigh!  Software development is hard.)

Signed-off-by: Eric Myhre <hash@exultant.us>