Fixup on docs strings on Node methods.
These have been just slightly out of style for a while now.

(In particular, allusions to numeric-keyed maps are removed.
While these are still considered a potential "someday" subject rather
than fully ruled out, it's also certainly not a *present* feature;
until something changes on this front, better not to advertise it.)

Signed-off-by: Eric Myhre <hash@exultant.us>