Add TypeName field to ErrWrongKind.
In general, picking a consistent strategy with this error and sticking
with it.  The strings were starting to accumulate too much ~stuff~,
otherwise.

It might not be a bad idea to upgrade the MethodName strings to a bunch
of consts, but I don't think it will get any harder to do that in the
future if we should so choose either, so, meh.

Signed-off-by: Eric Myhre <hash@exultant.us>