whitespace nuzzling.
This is pretty much a lost cause, though.  If we're pretending these
methods can be called in any order, then getting an outcome that's
guaranteed to match what go-fmt will do is not possible without also
passing more context around than we are.

So, I'm just running on a loose heuristic of clustering similar things
together, and trying roughly to get a linebreak between sections that
have a bigger jump between them.  It's... eh.  Not gonna spend much
time on it.

Signed-off-by: Eric Myhre <hash@exultant.us>