Fix two subtle errors in error text generation.
It always bamboozles me that ranging modifies the "dot" in go
templates.  Even if you're assigning the range yield into other values
explicitly: it still frobs dot.  Why.

In this case, it was also a silent error because both the template
input object and a type field both have ".Type.Name" properties.  Ow.

Signed-off-by: Eric Myhre <hash@exultant.us>