Add accessors to struct representation details.
Will need this in a moment for the body of the representation codegen
for structs-represented-as-maps.

This is not a particularly strongly consistency-checked method; if you
give it a struct field from another struct, it'll silently do a wrong
thing.  This is because checking for that would require back-pointers,
and I... don't wanna.  Can revisit this if it becomes problematic.

Signed-off-by: Eric Myhre <hash@exultant.us>