Doc for expected error behavior on typed nodes.
It does vary from the baselines somewhat.

There's a fixme in the codegen for struct-with-repr-map for where to
address the bug found in the test commit preceeding this one, but
solving it well depends on another change I've been planning in core,
so I'm probably going to park this branch soon and go do that work
on master; then come back to fix this here after that lands.

Signed-off-by: Eric Myhre <hash@exultant.us>