More comments on map impl options.  Partial decisions.
We really, truly cannot get away from having a KeyAssembler option.
That much seems clear.  So that's moved up into the interface we're
working with, and one of the alternate drafts dropped.

The comments around the NodeBuilder.Reset method are resolved: the
parts about building keys are removed, as we've concluded pretty
solidly that those were indeed about to take a wrong turn.
(Resetting a builder is still a useful thing overall if you're going
to make a bunch of nodes of a single type: the node allocations
will still happen, but the builder being resettable should be
pretty much a giant duff's device action, and saves one of the
two mega-amortized allocations in a codegen'd world, which is better
than not.)

Dropped the "demo" function.  This isn't the place for it, and that
demo had also gone awry.

Some new questions about some of the other "simple" methods on
MapAssembler and ListAssembler.  Are they useful shortcuts (that
actually provide a performance benefit if present directly on the
assembler interface)?  Or are they trivially recreateable from a
freestanding function operating the interface, and thus better be
done that way, to save SLOC in codegen outputs?

I think I'm going to start needing benchmarks to come to useful
conclusions on that last question pair, so... it's time to move forward
on writing some more implementation code matching these interfaces.
(Unfortunately, I think there's at least... four different
implementations needed to get a good eye on things, since the codegen
path vs runtime fully generic interfaces path have *very* different
characteristics, and then we need each of the styles of interest
implemented for each.  But that's just what we'll have to do.)